All Environ objects are upgraded to the new structure, optimized. Added EnvironObject base class. Menu state is updated to use grass object. CMakeList.txt is cleaned and structurized now.

+++ include/World/WorldManager.hpp
@@ -59,7 +59,7 @@ class WorldManager : public boost::serialization::singleton<WorldManager>
         float getWindPower();
 
         void setDropLimit(float maxDrops);
-        float getDropLimit();
+        int getDropLimit();
 
         float getMoonAngle();
